refactored, moved like button to its own component
