+++ frontend/src/components/home/card__like_btn/card__like_btn.css
@@ -1,0 +1,3 @@
+.card__like_btn {
+	
+}
